+++ NEWS.md
@@ -1,3 +1,11 @@
+# pedquant 0.1.5
+
+* fixed a bug caused by Chinese characters in pq_plot function
+* fixed a bug in mk_stock skip index for adjfactor type
+* fixed the date of spot price from 163
+* update the Chinese district via admin_mca function
+* md_stock supports to load the data of fund in sse/szse
+
 # pedquant 0.1.3
 
 * Fixed a bug in ed_nbs when the symbol provided is wrong.